tests/submission/Opera/preload/metadata/preload-metadata-to-none-after-src-networkstate-dataurl.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-to-none-after-src-networkstate.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata networkState when preload is set to none after src is set - wave</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata networkState when preload is set to none after src is set - wave", "HTMLVideoElement.preload:metadata networkState when preload is set to none after src is set - wave");
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var states_actual = '';
	var node = document.getElementsByTagName(tag_name)[0];
	node.addEventListener('suspend', endTest, false);
	node.addEventListener('error', function() {
	    if (!node.canPlayType(tag_name + '/wave')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/wave'");
		    t.done();
		});
	    }
	}, false);
	node.src = 'data:audio/wav;base64,UklGRigAAABXQVZFZm10IBAAAAABAAEARKwAAIhYAQACABAAZGF0YQQAAAAAAAAA';
	node.preload = 'none';

	function endTest() {
	    t.step(function() {
		assert_equals(node.networkState, HTMLMediaElement.NETWORK_IDLE, 'networkState after suspend')
		t.done();
	    });
	}
    });
});
</script>
